MyGameInstance created a game instance as a carrier for global variables: GI_xval GI_yval GI_zval which carry the respective values of which column's index to pull the coordinate data from no duplicate control is in place, e.g. the same column can be selected to use as the X, Y, and Z values (e.g. a linear graph albeit in 3D) TO DO: rename the variable so that it indicates that it contains the index of the column to look for the value, not the value itself add an array of the column names contained in the data table and the structure that the data table is loaded from. this will allow the creation of other elements, such as the drop-down boxes in the menu, to be done dynamically, which is good coding form
ScenicMenu
Created a second level that houses the menu and spawns a widget from a blueprint. it then partially populates the
menu with options
TO DO: finish the population of the drop-down boxes that are present in the blueprint, design and add a
functionality for the checkboxes

MainMenuWB
created a widget blueprint that contains drop-down boxes, checkboxes, that allow the user to select which data
values to use as the coordinate values of the dots. two buttons allow the user to control the order of play.

MainLevel
Event graph now grabs the X, Y, Z index values from the global variables in the new game instance, and uses them
to determine which values to pass on to the suburb blueprint when spawning actors.
TO DO: create a compass rose, based off the global variables

COPY
created a new material by copying the default material used for rendering text, and aptly named it "copy".
This material has the emmissive property, so it will display its colours in any lighting
TO DO:
give the material a more intuitive name

Suburb
The Blueprint contains 4 new variables:
Name - contains the name/title/descriptor of the data point
Xval, Yval, Zval - contains the individual coordinate value that the suburb actor is to be placed at

The generation of the position vector is now done using the new X, Y, Z variables' values.
Text is now rendered next to each suburb actor, and contains the Name variable's value
The text is rendered using the new material "COPY", detailed above, so the text is always visible
each time a suburb actor is spawned, it gets a random colour value, to differentiate it.
TO DO:
Have the colour of each actor's mesh/textrender/both be controlled by some variable, possibly set in the main menu